Add integration tests.
